requirements.txt
-i https://pypi.org/simple
adal==1.2.5
asgiref==3.2.10; python_version >= '3.5'
astroid==2.4.1; python_version >= '3.5'
attrs==20.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
azure-common==1.1.25
azure-core==1.8.2
azure-mgmt-core==1.2.1
azure-mgmt-media==1.1.1
azure-mgmt-resource==15.0.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
decorator==4.4.2
django-appconf==1.0.4
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-extensions==3.0.9
django-imagekit==4.0.2
django-materializecss-form==1.1.17
django-registration-redux==2.8
django-restframework==0.0.1
django-storages[azure]==1.10.1
django-taggit==1.3.0
django==3.1.2
dodgy==0.2.1
flake8-polyfill==1.0.2
flake8==3.8.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gunicorn==20.0.4
h11==0.11.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
imageio-ffmpeg==0.4.2; python_version >= '3.4'
imageio==2.9.0
iniconfig==1.1.1
ipython-genutils==0.2.0
ipython==7.18.1
isodate==0.6.0
isort==4.3.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jedi==0.17.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
lazy-object-proxy==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
mccabe==0.6.1
moviepy==1.0.3
msrest==0.6.19
msrestazure==0.6.4
numpy==1.19.2; python_version >= '3.6'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pep8-naming==0.10.0
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pilkit==2.0
pillow==8.0.1
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
proglog==0.1.9
prompt-toolkit==3.0.8; python_full_version >= '3.6.1'
prospector==1.3.1
psycopg2-binary==2.8.6
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycodestyle==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydocstyle==5.1.1; python_version >= '3.5'
pyflakes==2.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.7.2; python_version >= '3.5'
pyjwt==1.7.1
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pylint==2.5.3; python_version >= '3.5'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
pytest-django==4.1.0
pytest==6.1.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2020.1
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requirements-detector==0.7
setoptconf==0.2.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
snowballstemmer==2.0.0
sqlparse==0.4.1; python_version >= '3.5'
toml==0.10.1
tqdm==4.51.0; python_version >= '2.6' and python_version not in '3.0, 3.1'
traitlets==5.0.5; python_version >= '3.7'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.12.2
wcwidth==0.2.5
werkzeug==1.0.1
wrapt==1.12.1
